7project/project/backend/requirements.txt
flask==2.3.3
mysql-connector-python==8.1.0
redis==4.6.0
psycopg2-binary #Para postgresql database